Matrix derivative computations edited
The current notation is a little confusing and in parts does not seem to agree with other sources (see http://argento.bu.edu/hes/MatrixCalculus.pdf). For example, how do you differentiate with respect to a matrix? Most texts define these derivatives with respect to vectors or scalars. Two unnecessary lines were removed and the proof of minimization was adapted to the corrected differentiation rules.